safari-534.53-branch/LayoutTests/fast/dom/document-attribute-js-null-expected.txt

This test setting various attributes of documents to JavaScript null.

TEST SUCCEEDED: Got the expected exception (9). [tested Document.xmlVersion]
TEST SUCCEEDED: The value was null. [tested Document.documentURI]
TEST SUCCEEDED: The value was undefined. [tested Document.charset]
TEST SUCCEEDED: The value was null. [tested Document.selectedStylesheetSet]

TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.title]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.cookie]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.bgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.fgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.alinkColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.linkColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.vlinkColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLDocument.dir]
TEST SUCCEEDED: The value was the string 'off'. [tested HTMLDocument.designMode]
